Factored quizzes out of main.
